refactor #287 : 모달 공통 스타일 및 타입 분리
