+++ src/common/components/modal/base-modal.tsx
@@ -1,26 +1,10 @@
-import { ReactNode } from 'react';
-
 import * as Dialog from '@radix-ui/react-dialog';
 
-import * as styles from './base-modal.styles';
+import { ModalContentProps, ModalProps } from './modal-types';
 
-interface BaseModalProps {
-  open: boolean;
-  onOpenChange: (open: boolean) => void;
-  children: ReactNode;
-}
+import * as styles from './index.styles';
 
-interface ModalContentProps {
-  children: ReactNode;
-  width?: string;
-  height?: string;
-  padding?: string;
-  backgroundColor?: string;
-  gap?: string;
-  zIndex: number;
-}
-
-function BaseModal({ open, onOpenChange, children }: BaseModalProps) {
+function BaseModal({ open, onOpenChange, children }: ModalProps) {
   if (!open) return null;
   return (
     <Dialog.Root open={open} onOpenChange={onOpenChange}>
